[#936] Extract colorant-family tags into tags/colorant.py (item 4)
Created `DisplayCAL/icc_profile/tags/colorant.py` with `Colorant`,
`ColorantTableType`, `ChromaticityType`, `Geometry`, `Illuminant`,
`Observer`, `MeasurementType`, `ViewingConditionsType`, and
`ProfileSequenceDescType`, moved out of `icc_profile/__init__.py`.

`ProfileSequenceDescType.__init__()` and `.add()` need a deferred import
for `TextDescriptionType`/`MultiLocalizedUnicodeType`, which still live
in `icc_profile/__init__.py` pending their own extraction into
`tags/text.py` in item 5, mirroring the pattern already used for
`_mp_apply_black`/`pool_slice` in `LUT16Type`.

Also dropped `COLORANTS`/`GEOMETRY`/`ILLUMINANTS`/`OBSERVERS` and
`collections.abc.Iterator`, which became genuinely unused in
`__init__.py` after this extraction (verified no external consumer
imports these constants from `icc_profile` specifically, only from
their real home modules).
